update arm-ttk: Parameter Types Should Be Consistent
Type Mismatch: Parameter 'wlsIdentityKeyStoreData' in nested tem
plate 'wls-domain-deployment' is defined as string, but the parent template def
ines it as secureString). Line: 4883, Column: 22
        Type Mismatch: Parameter 'wlsPrivateKeyAlias' in nested temp
late 'wls-domain-deployment' is defined as string, but the parent template defi
nes it as secureString). Line: 4914, Column: 22
        Type Mismatch: Parameter 'wlsTrustKeyStoreData' in nested te
mplate 'wls-domain-deployment' is defined as string, but the parent template de
fines it as secureString). Line: 4922, Column: 22
        Type Mismatch: Parameter 'wlsIdentityKeyStoreData' in nested
 template 'wls-domain-deployment' is defined as string, but the parent template
 defines it as secureString). Line: 3330, Column: 22
        Type Mismatch: Parameter 'wlsPrivateKeyAlias' in nested temp
late 'wls-domain-deployment' is defined as string, but the parent template defi
nes it as secureString). Line: 3361, Column: 22
        Type Mismatch: Parameter 'wlsTrustKeyStoreData' in nested te
mplate 'wls-domain-deployment' is defined as string, but the parent template de
fines it as secureString). Line: 3369, Column: 22
        Type Mismatch: Parameter 'wlsIdentityKeyStoreData' in nested
 template 'setup-wls-cluster' is defined as secureString, but the parent templa
te defines it as string). Line: 2511, Column: 14
        Type Mismatch: Parameter 'wlsPrivateKeyAlias' in nested temp
late 'setup-wls-cluster' is defined as secureString, but the parent template de
fines it as string). Line: 2548, Column: 14
        Type Mismatch: Parameter 'wlsTrustKeyStoreData' in nested te
mplate 'setup-wls-cluster' is defined as secureString, but the parent template
defines it as string). Line: 2556, Column: 14Type Mismatch: Parameter 'wlsIdentityKeyStoreData' in nested tem
plate 'wls-domain-deployment' is defined as string, but the parent template def
ines it as secureString). Line: 4883, Column: 22
        Type Mismatch: Parameter 'wlsPrivateKeyAlias' in nested temp
late 'wls-domain-deployment' is defined as string, but the parent template defi
nes it as secureString). Line: 4914, Column: 22
        Type Mismatch: Parameter 'wlsTrustKeyStoreData' in nested te
mplate 'wls-domain-deployment' is defined as string, but the parent template de
fines it as secureString). Line: 4922, Column: 22
        Type Mismatch: Parameter 'wlsIdentityKeyStoreData' in nested
 template 'wls-domain-deployment' is defined as string, but the parent template
 defines it as secureString). Line: 3330, Column: 22
        Type Mismatch: Parameter 'wlsPrivateKeyAlias' in nested temp
late 'wls-domain-deployment' is defined as string, but the parent template defi
nes it as secureString). Line: 3361, Column: 22
        Type Mismatch: Parameter 'wlsTrustKeyStoreData' in nested te
mplate 'wls-domain-deployment' is defined as string, but the parent template de
fines it as secureString). Line: 3369, Column: 22
        Type Mismatch: Parameter 'wlsIdentityKeyStoreData' in nested
 template 'setup-wls-cluster' is defined as secureString, but the parent templa
te defines it as string). Line: 2511, Column: 14
        Type Mismatch: Parameter 'wlsPrivateKeyAlias' in nested temp
late 'setup-wls-cluster' is defined as secureString, but the parent template de
fines it as string). Line: 2548, Column: 14
        Type Mismatch: Parameter 'wlsTrustKeyStoreData' in nested te
mplate 'setup-wls-cluster' is defined as secureString, but the parent template
defines it as string). Line: 2556, Column: 14

Signed-off-by: galiacheng <haixia.cheng@microsoft.com>

 Changes to be committed:
	modified:   src/main/bicep/modules/_azure-resoruces/_keyvault/_keyvaultForWLSSSLCert.bicep
	modified:   src/main/bicep/modules/_deployment-scripts/_ds-create-wls-cluster.bicep